+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/RandomWrapperSecureRandomTest.java
@@ -1,5 +1,7 @@
 package io.github.pr0methean.betterrandom.prng;
 
+import static org.testng.Assert.assertEquals;
+
 import io.github.pr0methean.betterrandom.seed.DefaultSeedGenerator;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import java.security.GeneralSecurityException;
@@ -7,6 +9,8 @@ import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.util.Map;
 import java.util.Random;
+import java.util.concurrent.ConcurrentSkipListSet;
+import java.util.function.Function;
 import org.testng.annotations.Test;
 
 public class RandomWrapperSecureRandomTest extends BaseRandomTest {
@@ -76,5 +80,10 @@ public class RandomWrapperSecureRandomTest extends BaseRandomTest {
     final RandomWrapper wrapper = createRngInternal();
     wrapper.setSeed(seed);
     return wrapper;
+  }
+
+  /** Assertion-free because SecureRandom isn't necessarily reproducible. */
+  @Override @Test public void testThreadSafety() {
+    testThreadSafetyVsCrashesOnly(FUNCTIONS_FOR_THREAD_SAFETY_TEST);
   }
 }