+++ mutation.sh
@@ -2,5 +2,16 @@
 cd betterrandom
 unset RANDOM_DOT_ORG_KEY # Would probably exhaust even our 5Mbit/day limit
 mvn compile test-compile org.pitest:pitest-maven:mutationCoverage
-cd ..
-git clone https://github.com/Pr0methean/pr0methean.github.io.git docs
+cd ../docs
+git remote add originauth "https://${GH_TOKEN}@github.com/Pr0methean/pr0methean.github.io.git"
+git pull --rebase originauth master
+git checkout originauth/master
+rm -rf docs/betterrandom-pit-reports
+mv ../betterrandom/target/pit-reports betterrandom-java7-pit-reports
+git add betterrandom-pit-reports
+git commit -m "Update PIT mutation reports"
+git push originauth HEAD:master
+while [ ! $? ]; do
+  git pull --rebase # Merge
+  git push
+done